requirements.txt
discord.py==2.3.2
anthropic>=0.40.0
python-dotenv==1.0.0
certifi
audioop-lts
